+++ src/Normalizer/FromArray/ContactInfoNormalizer.php
@@ -6,8 +6,11 @@ namespace DigipolisGent\Toerismevlaanderen\Lodging\Normalizer\FromArray;
 
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\ContactInfo;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddresses;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumber;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumbers;
 use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddress;
+use DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddresses;
 
 /**
  * Normalizes an array of data into a ContactInfo value.
@@ -23,16 +26,61 @@ final class ContactInfoNormalizer
      */
     public function normalize(array $data): ContactInfo
     {
-        $phoneNumber = !empty($data['phoneNumber'])
-            ? PhoneNumber::fromNumber($data['phoneNumber'])
-            : PhoneNumber::withoutNumber();
-        $emailAddress = !empty($data['emailAddress'])
-            ? EmailAddress::fromAddress($data['emailAddress'])
-            : EmailAddress::withoutAddress();
-        $websiteAddress = !empty($data['websiteAddress'])
-            ? WebsiteAddress::fromUrl($data['websiteAddress'])
-            : WebsiteAddress::withoutUrl();
+        return ContactInfo::fromDetails(
+            $this->normalizePhoneNumbers($data['phoneNumbers'] ?? []),
+            $this->normalizeEmailAddresses($data['emailAddresses'] ?? []),
+            $this->normalizeWebsiteAddresses($data['websiteAddresses'] ?? [])
+        );
+    }
 
-        return ContactInfo::fromDetails($phoneNumber, $emailAddress, $websiteAddress);
+    /**
+     * Normalize the phone numbers.
+     *
+     * @param string[] $numbers
+     *
+     * @return \DigipolisGent\Toerismevlaanderen\Lodging\Value\PhoneNumbers
+     */
+    private function normalizePhoneNumbers(array $numbers): PhoneNumbers
+    {
+        $phoneNumbers = [];
+        foreach ($numbers as $number) {
+            $phoneNumbers[] = PhoneNumber::fromNumber($number);
+        }
+
+        return PhoneNumbers::fromPhoneNumbers(...$phoneNumbers);
+    }
+
+    /**
+     * Normalize the email addresses.
+     *
+     * @param string[] $addresses
+     *
+     * @return \DigipolisGent\Toerismevlaanderen\Lodging\Value\EmailAddresses
+     */
+    private function normalizeEmailAddresses(array $addresses): EmailAddresses
+    {
+        $emailAddresses = [];
+        foreach ($addresses as $address) {
+            $emailAddresses[] = EmailAddress::fromAddress($address);
+        }
+
+        return EmailAddresses::fromEmailAddresses(...$emailAddresses);
+    }
+
+    /**
+     * Normalize the website addresses.
+     *
+     * @param string[] $urls
+     *
+     * @return \DigipolisGent\Toerismevlaanderen\Lodging\Value\WebsiteAddresses
+     */
+    private function normalizeWebsiteAddresses(array $urls): WebsiteAddresses
+    {
+        $websiteAddresses = [];
+        foreach ($urls as $url) {
+            $websiteAddresses[] = WebsiteAddress::fromUrl($url);
+        }
+
+        return WebsiteAddresses::fromWebsiteAddresses(...$websiteAddresses);
     }
 }